return to the first version
Some long time ago I posted this feature request. 

Well, I've recently started development on this for fun (and because I was having problems with some other code). 

The attached file allows for rotating the whole desktop (more or less) around one axis of the Rubik's cube. It also allows for the number of strips. 

Also, it only works for a horizontal desktop size of 4 at the moment but I'm pretty sure it could work for any hsize. 

Finally, I intend for this to be incorporated into the screensaver plugin or act much like a screensaver and don't know how fast development will be on this plugin.